back/requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.1.2
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
soupsieve==2.6
threadpoolctl==3.5.0
urllib3==2.2.3
Werkzeug==3.0.6
